# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CHELSY Restaurant API",
        "description": "API REST pour l'application mobile CHELSY Restaurant",
        "contact": {
            "email": "contact@chelsy-restaurant.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Liste des commandes de l'utilisateur",
                "operationId": "00d3856d2d3d93124e0cc021d0cb5b79",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commandes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Créer une nouvelle commande",
                "operationId": "79da0a1adcc617490783dfd61927ff56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "payment_method"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "delivery",
                                            "pickup"
                                        ],
                                        "example": "delivery"
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "card",
                                            "cash",
                                            "mobile_money"
                                        ],
                                        "example": "card"
                                    },
                                    "mobile_money_provider": {
                                        "type": "string",
                                        "enum": [
                                            "MTN",
                                            "Moov"
                                        ],
                                        "nullable": true
                                    },
                                    "mobile_money_number": {
                                        "type": "string",
                                        "example": "+229 12 34 56 78",
                                        "nullable": true
                                    },
                                    "promo_code": {
                                        "type": "string",
                                        "example": "PROMO10",
                                        "nullable": true
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "datetime",
                                        "nullable": true
                                    },
                                    "special_instructions": {
                                        "type": "string",
                                        "example": "Sonner 2 fois",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commande créée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "type": "object"
                                                },
                                                "payment": {
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Détails d'une commande",
                "operationId": "b2f9845bf9fe11158e0e43f1dd0c5802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/orders/{id}/invoice/download": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Télécharger la facture PDF",
                "operationId": "5095a7592388440a7f2627695a4bd3bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF de la facture",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}